Move dom manipulation to accountPageUtils. Add SkeletonLoader
