+++ src/client/styles/components/AccountPage.scss
@@ -23,3 +23,7 @@
     }
   }
 }
+
+.loading#account-page-content {
+  display: none;
+}
